add ft_ping beta

+++ ft_ping/src/main.c
@@ -6,7 +6,7 @@
 /*   By: mbougrin <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/18 11:02:44 by mbougrin          #+#    #+#             */
-/*   Updated: 2016/10/26 12:13:18 by mbougrin         ###   ########.fr       */
+/*   Updated: 2016/10/26 12:14:08 by mbougrin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -214,8 +214,8 @@ void					printSigint(void)
 {
 	t_stc 			*stc = singleton(NULL);
 	int				ret;
-	double			one;
-	double 			tmp;
+	float			one;
+	float 			tmp;
 
 	ret = 100;
 	one = stc->count / 100;
